+++ pyconfix/__init__.py
@@ -1,3 +1,5 @@
+from ._version import __version__
+
 from .runner import pyconfix
 from .core import Core
 from .option import ConfigOption, ConfigOptionType